requirements.txt
# Immigration Location Analysis - Requirements
# Python 3.12+ required

# Core dependencies
pandas>=2.2.0
numpy>=1.26.0
pyyaml>=6.0.1

# Geographic calculations
geopy>=2.4.1

# Visualization - Static
matplotlib>=3.8.0
seaborn>=0.13.0

# Visualization - Interactive
plotly>=5.18.0
folium>=0.15.1

# Utilities
python-dateutil>=2.8.2

# Optional (for development)
# pytest>=7.4.0
# black>=23.12.0
# ruff>=0.1.9
# mypy>=1.7.0
